requirements.txt
flask
pandas
requests
beautifulsoup4
python-dotenv
selenium